AndroidDebugBridge: move DeviceMonitor-related events to the DeviceMonitor class.
